+++ profiles/tables/__init__.py
@@ -4,3 +4,4 @@ from .role_table import *
 from .team_table import *
 from .user_table import *
 from .permission_table import *
+from .approval_workflow import *